pyproject.toml
[tool.poetry]
name = "course-work-django"
version = "0.1.0"
description = ""
authors = ["DalOpez <alex-grazhdankin@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.1.0"
pillow = "^11.2.1"
django-countries = "^7.6.1"
ipython = "^9.2.0"
black = "^25.1.0"
redis = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
